+++ Catalogo.Application/Services/ProdutoService.cs
@@ -1,6 +1,7 @@
 ﻿using AutoMapper;
 using Catalogo.Application.DTOs;
 using Catalogo.Application.Interfaces;
+using Catalogo.Domain.Entities;
 using Catalogo.Domain.Interfaces;
 using System;
 using System.Collections.Generic;
@@ -37,22 +38,61 @@ public class ProdutoService : IProdutoService
 
     public async Task Add(ProdutoDTO produtoDto)
     {
-        var produtoEntity = _mapper.Map<Domain.Entities.Produto>(produtoDto);
+        var produtoEntity = _mapper.Map<Produto>(produtoDto);
 
         await _produtoRepository.CreateAsync(produtoEntity);
     }
 
-    public async Task Update(ProdutoDTO produtoDto)
+    public async Task<ProdutoDTO> Update(ProdutoDTO produtoDTO)
     {
-        var produtoEntity = _mapper.Map<Domain.Entities.Produto>(produtoDto);
+        var produtoEntity = await _produtoRepository.GetByIdAsync(produtoDTO.Id);
+
+        if (produtoEntity is null)
+            return null;
+
+        produtoEntity.Update(produtoDTO.Nome, produtoDTO.Descricao, produtoDTO.Preco, produtoDTO.Imagemurl, 
+            produtoDTO.Estoque, produtoDTO.DataCadastro, produtoDTO.CategoriaId);
 
         await _produtoRepository.UpdateAsync(produtoEntity);
+
+        return produtoDTO;
     }
 
-    public async Task Remove(int id)
+    public async Task<ProdutoDTO> Remove(int id)
     {
         var produtoEntity = await _produtoRepository.GetByIdAsync(id);
 
+        if (produtoEntity is null)
+            return null;
+
         await _produtoRepository.RemoveAsync(produtoEntity);
+
+        return _mapper.Map<ProdutoDTO>(produtoEntity);
+    }
+
+    public async Task<ProdutoDTO> Patch(int id, ProdutoPatchDTO produtoPatchDTO)
+    {
+        var produtoEntity = await _produtoRepository.GetByIdAsync(id);
+
+        produtoEntity = AplicarAlteracoes(produtoEntity, produtoPatchDTO);
+
+        await _produtoRepository.UpdateAsync(produtoEntity);
+
+        return _mapper.Map<ProdutoDTO>(produtoEntity);
+    }
+
+    // métodos auxiliares
+
+    private Produto AplicarAlteracoes(Produto produtoEntity, ProdutoPatchDTO produtoPatchDTO)
+    {
+        string novoNome = produtoPatchDTO.Nome ?? produtoEntity.Nome;
+        string novaDescricao = produtoPatchDTO.Descricao ?? produtoEntity.Descricao;
+        string novaImagemUrl = produtoPatchDTO.Imagemurl ?? produtoEntity.Imagemurl;
+        decimal novoPreco = produtoPatchDTO.Preco != null ? produtoPatchDTO.Preco : produtoEntity.Preco;
+        int novoEstoque = produtoPatchDTO.Estoque != null ? produtoPatchDTO.Estoque : produtoEntity.Estoque;
+
+        produtoEntity.Update(novoNome, novaDescricao, novoPreco, novaImagemUrl, novoEstoque, produtoEntity.DataCadastro, produtoEntity.CategoriaId);
+
+        return produtoEntity;
     }
 }
